+++ lib/components/EmailInput.js
@@ -2,45 +2,12 @@ import React, { useState } from "react";
 
 import classnames from "classnames";
 import PropTypes from "prop-types";
-import { components } from "react-select";
 import CreatableSelect from "react-select/creatable";
 
-import Label from "./Label";
+import { EMAIL_SEPARATION_REGEX, CUSTOM_STYLES, CUSTOM_COMPONENTS, formatEmailInputOptions } from "../constants/EmailInput";
+import { hyphenize } from "../common";
 import Typography from "./Typography";
-
-const CustomControl = ({ children, ...props }) => (
-  <components.Control {...props}>
-    <div className="overflow-hidden">{children}</div>
-  </components.Control>
-);
-
-const customComponents = {
-  DropdownIndicator: null,
-  ClearIndicator: null,
-  Control: CustomControl,
-};
-
-const isValidEmail = email => {
-  const regex = new RegExp("^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}$", "i");
-  return regex.test(email);
-};
-
-const formatEmailInputOptions = label => ({
-  label,
-  value: label,
-  valid: isValidEmail(label),
-});
-
-const customStyles = {
-  input: styles => ({
-    ...styles,
-    overflow: "hidden",
-  }),
-  multiValue: (styles, { data: { valid } }) => ({
-    ...styles,
-    border: valid ? "none" : "thin dashed #f56a58 !important",
-  }),
-};
+import Label from "./Label";
 
 const EmailInput = ({
   label = "",
@@ -54,9 +21,9 @@ const EmailInput = ({
   const [inputValue, setInputValue] = useState("");
 
   const handleEmailChange = () => {
-    const emails = inputValue.match(/[^\s,]+/g);
-    const values = emails.map(email => formatEmailInputOptions(email));
-    onChange([...value, ...values]);
+    const inputValues = inputValue.match(EMAIL_SEPARATION_REGEX);
+    const emails = inputValues.map(email => formatEmailInputOptions(email));
+    onChange([...value, ...emails]);
     setInputValue("");
   };
 
@@ -73,12 +40,12 @@ const EmailInput = ({
     }
   };
 
-  const handleBlur = event => inputValue ?  handleEmailChange() : onBlur(event);
+  const handleBlur = event => inputValue ? handleEmailChange() : onBlur(event);
 
   return (
     <div className="flex flex-col">
       {label && (
-        <Label className="mb-1" data-cy="emails-input-label">
+        <Label className="mb-1" data-cy={`${hyphenize(label)}-input-label`}>
           {label}
         </Label>
       )}
@@ -86,7 +53,8 @@ const EmailInput = ({
         isMulti
         menuIsOpen={false}
         inputValue={inputValue}
-        components={customComponents}
+        components={CUSTOM_COMPONENTS}
+        data-cy={`${hyphenize(label)}-email-input`}
         classNamePrefix="neeto-ui-react-select"
         className={classnames(
           "neeto-ui-react-select__container",
@@ -98,7 +66,7 @@ const EmailInput = ({
         onBlur={handleBlur}
         placeholder={placeholder}
         value={value}
-        styles={customStyles}
+        styles={CUSTOM_STYLES}
         isDisabled={disabled}
       />
       <div>
